Provides more detailed permission verification and assignment file upload and download (#65)
