+++ .gitignore
@@ -9,3 +9,5 @@
 
 /catfood/.vscode
 /restart.sh
+/catfood/.idea/
+
