+++ catfood/experiment/utils.py
@@ -1,3 +1,6 @@
+import datetime
+
+
 def generate_response(data, isSuccess):
     return {
         "isSuccess": isSuccess,
@@ -14,7 +17,7 @@ def dict_filter(old_dict, keys):
 
 def student_assignments_filter(assignments_list):
     key_white_list = ['course_case_id',
-                      'submission_uploader', 'submission_file_token', 'submission_case_id']
+                      'submission_uploader', 'submission_case_id']
     for index, assignment in enumerate(assignments_list):
         if not assignment['submission_is_public']:
             assignments_list[index] = dict_filter(assignment, key_white_list)
@@ -22,19 +25,16 @@ def student_assignments_filter(assignments_list):
 
 
 def is_submission_valid(submission):
-    key_white_list = ['course_case_id',
-                      'submission_uploader', 'submission_file_token']
-    for key in submission.keys():
-        if key not in key_white_list:
-            return False
-    return True
+    valid_keys_set = {'course_case_id',
+                      'submission_file_name', 'course_id'}
+    return set(submission.keys()) == valid_keys_set
 
 
 def is_submission_retrieve_valid(old_submission, new_submission):
     if old_submission['submission_is_public']:
         return False
     key_white_list = ['course_case_id',
-                      'submission_uploader', 'submission_file_token', 'submission_case_id']
+                      'submission_uploader', 'submission_file_name', 'submission_case_id']
     for key in new_submission.keys():
         if key not in key_white_list:
             return False
@@ -45,3 +45,8 @@ def is_submission_delete_valid(old_submission):
     if old_submission['submission_is_public']:
         return False
     return True
+
+
+def check_ddl(end_time):
+    ddl = datetime.datetime.strptime(end_time.split('+')[0], '%Y-%m-%dT%H:%M:%S')
+    return datetime.datetime.now() < ddl
